Desktop/InfoMary-main/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
chainlit==1.3.1
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
fastapi==0.112.4
fastjsonschema==2.20.0
filetype==1.2.0
Flask==3.0.1
frozenlist==1.5.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.1.1
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.4
langsmith==0.0.87
Lazify==0.4.0
literalai==0.0.623
MarkupSafe==3.0.2
marshmallow==3.23.0
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.9.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==23.2
pandocfilters==1.5.1
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.37.2
syncer==2.0.3
tenacity==8.5.0
tiktoken==0.5.2
tinycss2==1.4.0
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.27.0
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.16.0
wsproto==1.2.0
yarl==1.16.0
zipp==3.20.2

llama-index==0.9.38
chromadb==0.4.21
docx2txt==0.8
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
chainlit==1.3.1
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
fastapi==0.112.4
fastjsonschema==2.20.0
filetype==1.2.0
Flask==3.0.1
frozenlist==1.5.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.1.1
langchain-community==0.0.20
langchain-core==0.1.23
langchain-huggingface==0.1.0
langchain-openai==0.0.4
langsmith==0.0.87
Lazify==0.4.0
literalai==0.0.623
MarkupSafe==3.0.2
marshmallow==3.23.0
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.9.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==23.2
pandocfilters==1.5.1
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.10.1
python-multipart==0.0.9